+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   - Spaces in chain/command names are now officially disallowed. Those
     would have always been broken anyways, but now there is a nice message
     and the commands are excluded from the various listings
+  - Command alternatives provided by the "did-you-mean" feature no longer
+    display the full command list (help)
 
 ## v0.9.0 (2017-05-20)